# src/main/java/disenodesistemas/backendfunerariaapp/web/dto/response/BrandResponseDto.java
package disenodesistemas.backendfunerariaapp.web.dto.response;

import com.fasterxml.jackson.annotation.JsonInclude;

@JsonInclude(JsonInclude.Include.NON_DEFAULT)
public record BrandResponseDto(
    Long id,
    String name,
    String webPage
) { }