# src/main/resources/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Backend Funeraria App API
  version: 2.1.0
  description: |
    Contrato HTTP del monolito `backend-funeraria-app`.

    La API esta organizada por slices funcionales y documenta:
    - endpoints disponibles hoy en controllers reales
    - requests y responses con sus DTOs actuales
    - reglas de seguridad por JWT ligado a dispositivo
    - errores esperables por validacion, autorizacion y reglas de negocio

    Convenciones importantes:
    - Los endpoints autenticados requieren `Authorization: Bearer <token>`.
    - Los endpoints autenticados tambien requieren `X-Device-Id`.
    - `POST /api/v1/users/login` y `POST /api/v1/users/refresh` aceptan `Idempotency-Key`.
    - Todas las respuestas incluyen `X-Trace-Id`; si el cliente envia `X-Correlation-Id`, la API lo propaga.
    - Los errores de negocio y validacion siguen el modelo `ProblemDetail`.
servers:
  - url: http://localhost:8081
    description: Entorno local
...
paths:
...
  /api/v1/users/rol/{email}:
    post:
      tags: [Users]
      summary: Agregar o actualizar rol de un usuario
      description: Solo ADMIN. Devuelve el set de roles actualizado del usuario.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XDeviceId'
        - $ref: '#/components/parameters/EmailPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RolRequestDto'
      responses:
        '200':
          description: Roles actualizados.
          content:
            application/json:
              schema:
                type: array
                uniqueItems: true
                items:
                  $ref: '#/components/schemas/RolRequestDto'
        '400':
          $ref: '#/components/responses/ValidationResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
...
  /api/v1/categories:
    get:
      tags: [Catalogs]
      summary: Listar categorias
      description: Endpoint publico utilizado para formularios y catalogos.
      responses:
        '200':
          description: Lista de categorias.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryResponseDto'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
    post:
      tags: [Catalogs]
      summary: Crear categoria
      description: Solo ADMIN.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XDeviceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequestDto'
      responses:
        '201':
          description: Categoria creada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponseDto'
        '400':
          $ref: '#/components/responses/ValidationResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '409':
          $ref: '#/components/responses/ConflictResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /api/v1/categories/{id}:
    get:
      tags: [Catalogs]
      summary: Obtener categoria por id
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XDeviceId'
        - $ref: '#/components/parameters/LongIdPath'
      responses:
        '200':
          description: Categoria encontrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponseDto'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
    put:
      tags: [Catalogs]
      summary: Actualizar categoria
      description: Solo ADMIN.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XDeviceId'
        - $ref: '#/components/parameters/LongIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequestDto'
      responses:
        '200':
          description: Categoria actualizada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponseDto'
        '400':
          $ref: '#/components/responses/ValidationResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
    delete:
      tags: [Catalogs]
      summary: Eliminar categoria
      description: Solo ADMIN.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XDeviceId'
        - $ref: '#/components/parameters/LongIdPath'
      responses:
        '200':
          description: Categoria eliminada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatusModel'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /api/v1/cities:
    get:
      tags: [Catalogs]
      summary: Listar ciudades por provincia
      description: Endpoint publico. Requiere el query param `province_id`.
      parameters:
        - $ref: '#/components/parameters/ProvinceIdQuery'
      responses:
        '200':
          description: Ciudades de la provincia indicada.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CityResponseDto'
        '400':
          $ref: '#/components/responses/ValidationResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /api/v1/cities/{id}:
    get:
      tags: [Catalogs]
      summary: Obtener ciudad por id
      description: Requiere autenticacion.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XDeviceId'
        - $ref: '#/components/parameters/LongIdPath'
      responses:
        '200':
          description: Ciudad encontrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CityResponseDto'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
...
  /api/v1/deceased:
    get:
      tags: [Deceased]
      summary: Listar fallecidos
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XDeviceId'
      responses:
        '200':
          description: Lista de fallecidos.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeceasedResponseDto'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
    post:
      tags: [Deceased]
      summary: Registrar fallecido
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XDeviceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeceasedRequestDto'
      responses:
        '201':
          description: Fallecido registrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeceasedResponseDto'
        '400':
          $ref: '#/components/responses/ValidationResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '409':
          $ref: '#/components/responses/ConflictResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /api/v1/deceased/{dni}:
    get:
      tags: [Deceased]
      summary: Buscar fallecido por DNI
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XDeviceId'
        - $ref: '#/components/parameters/DniPath'
      responses:
        '200':
          description: Fallecido encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeceasedResponseDto'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
    put:
      tags: [Deceased]
      summary: Actualizar fallecido
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XDeviceId'
        - $ref: '#/components/parameters/DniPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeceasedRequestDto'
      responses:
        '200':
          description: Fallecido actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeceasedResponseDto'
        '400':
          $ref: '#/components/responses/ValidationResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '409':
          $ref: '#/components/responses/ConflictResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
    delete:
      tags: [Deceased]
      summary: Eliminar fallecido
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XDeviceId'
        - $ref: '#/components/parameters/DniPath'
      responses:
        '200':
          description: Operacion completada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatusModel'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
...
  /api/v1/items/category/{id}:
    get:
      tags: [Items]
      summary: Listar items por categoria
      description: Disponible para ADMIN y USER.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XDeviceId'
        - $ref: '#/components/parameters/LongIdPath'
      responses:
        '200':
          description: Items de la categoria indicada.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ItemResponseDto'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT de acceso emitido por `/api/v1/users/login`.
  parameters:
    XDeviceId:
      name: X-Device-Id
      in: header
      required: true
      description: Identificador estable del dispositivo o cliente que creo la sesion.
      schema:
        type: string
        example: web-chrome-gabriel
...
    DniPath:
      name: dni
      in: path
      required: true
      description: Documento nacional de identidad.
      schema:
        type: integer
        format: int32
        example: 35123456
    LongIdPath:
      name: id
      in: path
      required: true
      description: Identificador numerico.
      schema:
        type: integer
        format: int64
        example: 1
    EmailPath:
      name: email
      in: path
      required: true
      description: Email del usuario.
      schema:
        type: string
        format: email
        example: admin@funeraria.local
...
    ProvinceIdQuery:
      name: province_id
      in: query
      required: true
      description: Id de la provincia.
      schema:
        type: integer
        format: int64
        example: 6
...
  responses:
    ValidationResponse:
      description: Error de validacion de request.
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceIdHeader'
        X-Correlation-Id:
          $ref: '#/components/headers/CorrelationIdHeader'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ValidationProblem'
          example:
            type: about:blank
            title: Validacion fallida
            status: 400
            detail: La solicitud contiene datos invalidos.
            traceId: 4bf92f3577b34da6a3ce929d0e0e4736
            errors:
              - field: email
                message: El email es requerido.
    UnauthorizedResponse:
      description: Request sin credenciales validas, token invalido o dispositivo no coincidente.
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceIdHeader'
        X-Correlation-Id:
          $ref: '#/components/headers/CorrelationIdHeader'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: about:blank
            title: Error de aplicacion
            status: 401
            detail: El token no coincide con el dispositivo actual.
            code: security.jwt.device.mismatch
            traceId: 4bf92f3577b34da6a3ce929d0e0e4736
            correlationId: corr-request-0001
    ForbiddenResponse:
      description: El usuario autenticado no tiene permisos para ejecutar la operacion.
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceIdHeader'
        X-Correlation-Id:
          $ref: '#/components/headers/CorrelationIdHeader'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: about:blank
            title: Acceso denegado
            status: 403
            detail: Access Denied
            traceId: 4bf92f3577b34da6a3ce929d0e0e4736
    NotFoundResponse:
      description: Recurso no encontrado.
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceIdHeader'
        X-Correlation-Id:
          $ref: '#/components/headers/CorrelationIdHeader'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: about:blank
            title: Error de aplicacion
            status: 404
            detail: No existe afiliado con el DNI especificado.
            code: affiliate.error.not.found
            traceId: 4bf92f3577b34da6a3ce929d0e0e4736
    ConflictResponse:
      description: Conflicto de negocio, duplicado o estado incompatible.
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceIdHeader'
        X-Correlation-Id:
          $ref: '#/components/headers/CorrelationIdHeader'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: about:blank
            title: Error de aplicacion
            status: 409
            detail: Ya existe un afiliado registrado con el DNI especificado.
            code: affiliate.error.dni.already.exists
            traceId: 4bf92f3577b34da6a3ce929d0e0e4736
...
    InternalServerErrorResponse:
      description: Error interno del servidor.
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceIdHeader'
        X-Correlation-Id:
          $ref: '#/components/headers/CorrelationIdHeader'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: about:blank
            title: Error interno del servicio
            status: 500
            detail: Ocurrio un error interno inesperado.
            code: error.internal
            traceId: 4bf92f3577b34da6a3ce929d0e0e4736
  headers:
    TraceIdHeader:
      description: Identificador unico de la request generado o propagado por el servicio.
      schema:
        type: string
        example: 4bf92f3577b34da6a3ce929d0e0e4736
    CorrelationIdHeader:
      description: Identificador opcional provisto por el cliente para correlacionar requests entre sistemas.
      schema:
        type: string
        example: corr-request-0001
  schemas:
    ProblemDetail:
      type: object
      description: Modelo de error estandar usado por el backend.
      example:
        type: about:blank
        title: Error de aplicacion
        status: 404
        detail: No existe afiliado con el DNI especificado.
        code: affiliate.error.not.found
        traceId: 4bf92f3577b34da6a3ce929d0e0e4736
        correlationId: corr-request-0001
      properties:
        type:
          type: string
          format: uri
          nullable: true
          example: about:blank
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri
          nullable: true
        code:
          type: string
          description: Clave funcional vinculada a `messages_es.properties`.
        traceId:
          type: string
          description: Identificador unico de la request para correlacionar logs, errores y metricas.
          example: 4bf92f3577b34da6a3ce929d0e0e4736
        correlationId:
          type: string
          nullable: true
          description: Identificador opcional provisto por el cliente y propagado por el backend.
          example: corr-request-0001
        retryAfterSeconds:
          type: integer
          nullable: true
      additionalProperties: false
    ValidationProblem:
      example:
        type: about:blank
        title: Validacion fallida
        status: 400
        detail: La solicitud contiene datos invalidos.
        traceId: 4bf92f3577b34da6a3ce929d0e0e4736
        errors:
          - field: email
            message: El email es requerido.
      allOf:
        - $ref: '#/components/schemas/ProblemDetail'
        - type: object
          properties:
            errors:
              type: array
              items:
                $ref: '#/components/schemas/ValidationError'
    ValidationError:
      type: object
      properties:
        field:
          type: string
          example: email
        message:
          type: string
          example: no debe estar vacio
      required: [field, message]
...
    OperationStatusModel:
      type: object
      properties:
        name:
          type: string
          example: DELETE ITEM
        result:
          type: string
          example: SUCCESSFUL
      required: [name, result]
    UserDto:
      type: object
      properties:
        email:
          type: string
          format: email
          example: usuario@funeraria.local
        lastName:
          type: string
          example: Perez
        firstName:
          type: string
          example: Ana
      required: [email, lastName, firstName]
...
    ProvinceDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        code31662:
          type: string
          example: AR-B
        name:
          type: string
          example: Buenos Aires
    ProvinceResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        code31662:
          type: string
      required: [id, name]
    CityDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        province:
          $ref: '#/components/schemas/ProvinceDto'
        name:
          type: string
          example: La Plata
        zipCode:
          type: string
          example: 1900
    CityResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        zipCode:
          type: string
        province:
          $ref: '#/components/schemas/ProvinceResponseDto'
      required: [id, name]
    AddressRequestDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        apartment:
          type: string
          example: B
        blockStreet:
          type: integer
          example: 1234
        flat:
          type: string
          example: 4
        streetName:
          type: string
          example: Calle 50
        city:
          $ref: '#/components/schemas/CityDto'
      required: [streetName, city]
    AddressResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        streetName:
          type: string
        blockStreet:
          type: integer
        apartment:
          type: string
        flat:
          type: string
        city:
          $ref: '#/components/schemas/CityResponseDto'
...
    BrandResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        webPage:
          type: string
...
    CategoryRequestDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
      required: [name]
    CategoryResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
    GenderDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
      required: [id, name]
    GenderResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
      required: [id, name]
    RelationshipDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
      required: [id, name]
    RelationshipResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
      required: [id, name]
...
    DeathCauseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    DeathCauseResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
...
    RolRequestDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          enum: [ROLE_ADMIN, ROLE_USER]
      required: [name]
...
    DeceasedRequestDto:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        dni:
          type: integer
          format: int32
        birthDate:
          type: string
          format: date
          example: 1948-02-05
        deathDate:
          type: string
          format: date
          example: 2026-04-16
        placeOfDeath:
          $ref: '#/components/schemas/AddressRequestDto'
        gender:
          $ref: '#/components/schemas/GenderDto'
        deceasedRelationship:
          $ref: '#/components/schemas/RelationshipDto'
        deathCause:
          $ref: '#/components/schemas/DeathCauseDto'
        deceasedUser:
          $ref: '#/components/schemas/UserDto'
      required:
        [firstName, lastName, dni, birthDate, deathDate, gender, deceasedRelationship, deathCause]
    DeceasedResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        dni:
          type: integer
          format: int32
        affiliated:
          type: boolean
        birthDate:
          type: string
          description: Fecha con formato `dd-MM-yyyy`.
          example: 05-02-1948
        registerDate:
          type: string
          description: Fecha y hora con formato `dd-MM-yyyy HH:mm`.
          example: 17-04-2026 15:30
        deathDate:
          type: string
          description: Fecha con formato `dd-MM-yyyy`.
          example: 16-04-2026
        placeOfDeath:
          $ref: '#/components/schemas/AddressResponseDto'
        deceasedRelationship:
          $ref: '#/components/schemas/RelationshipResponseDto'
        deceasedUser:
          $ref: '#/components/schemas/UserDto'
        gender:
          $ref: '#/components/schemas/GenderResponseDto'
        deathCause:
          $ref: '#/components/schemas/DeathCauseResponseDto'
...
    ItemResponseDto:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        code:
          type: string
        itemImageLink:
          type: string
          example: https://storage.example/items/ataud-roble-01.png
        stock:
          type: integer
        price:
          type: number
        itemLength:
          type: number
        itemHeight:
          type: number
        itemWidth:
          type: number
        category:
          $ref: '#/components/schemas/CategoryResponseDto'
        brand:
          $ref: '#/components/schemas/BrandResponseDto'
...